% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_shift}
\alias{add_shift}
\title{Shift reference dates by a specified number of days}
\usage{
add_shift(df, n_day, refd_col)
}
\arguments{
\item{df}{A data frame containing a reference date column.}

\item{n_day}{An integer specifying the number of days to shift the reference date.
Positive values shift forward, negative values shift backward.}

\item{refd_col-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
A data frame with the reference date column shifted.
}
\description{
This function shifts the reference date column by a given number of days.
}
