% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_not_exist}
\alias{create_dir_not_exist}
\title{Create a directory and any missing parent directories}
\usage{
create_dir_not_exist(path)
}
\arguments{
\item{path}{String specifying a directory to create.}
}
\value{
\code{path}, invisibly.
}
\description{
Create a directory and any missing parent directories
}
