% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_name_pattern}
\alias{create_name_pattern}
\title{Create pattern to match input files of a given type and signal}
\usage{
create_name_pattern(indicator, signal, file_type = c("daily", "rollup"))
}
\arguments{
\item{file_type}{A string, either "daily" or "rollup"}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Create pattern to match input files of a given type and signal
}
