% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{data_filteration}
\alias{data_filteration}
\title{Filter Training and Testing Data Based on Lag}
\usage{
data_filteration(test_lag, data, lag_pad)
}
\arguments{
\item{test_lag}{Numeric or vector, the reference lag used for filtering.}

\item{data}{Data frame containing the lagged data to be filtered.}

\item{lag_pad}{Numeric or vector, specifying the lag padding. If a single value is provided,
it is applied symmetrically. If a vector of length two is provided, the first element is used
for left padding and the second for right padding.}
}
\description{
This function filters the input data based on the specified test lag and lag padding.
The lag padding can be either a single value (applied symmetrically) or a vector of two values,
specifying separate left and right padding.
}
