% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evaluate}
\alias{evaluate}
\title{Evaluation of the test results based on WIS score}
\usage{
evaluate(test_data, taus, response)
}
\arguments{
\item{test_data}{dataframe with a column containing the prediction results of
each requested quantile. Each row represents an update with certain
(reference_date, report_date, location) combination.}

\item{taus-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Evaluation of the test results based on WIS score
}
