% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_test_result}
\alias{export_test_result}
\title{Export test results. Merge all the locations and test lags together for one output file}
\usage{
export_test_result(
  test_data,
  coef_data,
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  gamma,
  training_end_date,
  training_days,
  value_type,
  export_dir
)
}
\arguments{
\item{test_data}{A data frame containing testing data.}

\item{indicator}{A string representing the forecasting indicator.}

\item{signal}{A string representing the forecast signal.}

\item{geo_level}{A string specifying the geographical level (e.g., state, county).}

\item{signal_suffix}{A string suffix for the signal name.}

\item{lambda}{A numeric vector specifying regularization parameters.}

\item{gamma}{A numeric vector specifying additional model parameters.}

\item{training_end_date}{The end date for training data.}

\item{value_type}{A string specifying the type of value being forecasted (e.g., "count").}

\item{train_data}{A data frame containing training data.}

\item{response}{A string specifying the response variable.}

\item{taus}{A numeric vector of quantiles for prediction intervals.}

\item{params_list}{A list of model parameters used for training and prediction.}

\item{geo}{A string specifying the geographical location for forecasting.}

\item{lp_solver}{A solver used for linear programming optimization in the model.}

\item{model_save_dir}{A string specifying the directory to save trained models.}

\item{training_start_date}{The start date for training data.}
}
\description{
Export test results. Merge all the locations and test lags together for one output file
}
