% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_missing_updates}
\alias{fill_missing_updates}
\title{Fill missing updates in a time series dataset with lagged values
Get pivot table, filling NANs. If there is no update on issue date D but
previous reports exist for issue date D_p < D, all the dates between
\link{D_p, D} are filled with with the reported value on date D_p. If there is
no update for any previous issue date, fill in with 0.
This function creates a complete reference-report date grid and fills in missing updates.
If an update is missing on a certain issue date but previous updates exist,
it fills gaps with the last available value. If no previous updates exist, it fills with \code{0}.}
\usage{
fill_missing_updates(
  df,
  value_col,
  refd_col,
  lag_col,
  temporal_resol = "daily",
  max_report_override = NULL
)
}
\arguments{
\item{df}{A data frame containing the dataset with reference and lagged dates.}

\item{value_col}{A string specifying the column name containing the values
to be filled and processed.}

\item{temporal_resol}{A string specifying the temporal resolution of the data.
Accepts \code{"daily"} (default) or \code{"weekly"} for processing in daily or weekly intervals.}

\item{refd_col-template}{\link{DESCRIPTION GOES HERE}}

\item{lag_col-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Fill missing updates in a time series dataset with lagged values
Get pivot table, filling NANs. If there is no update on issue date D but
previous reports exist for issue date D_p < D, all the dates between
\link{D_p, D} are filled with with the reported value on date D_p. If there is
no update for any previous issue date, fill in with 0.
This function creates a complete reference-report date grid and fills in missing updates.
If an update is missing on a certain issue date but previous updates exist,
it fills gaps with the last available value. If no previous updates exist, it fills with \code{0}.
}
