% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_rows}
\alias{fill_rows}
\title{Re-index, fill na, make sure all reference date have enough rows for updates}
\usage{
fill_rows(df, refd_col, lag_col, min_refd, max_refd, ref_lag)
}
\arguments{
\item{df}{A data.frame}

\item{refd_col}{A string specifying the column name representing the
reference date (i.e., the date the report is issued for).}

\item{lag_col}{A string specifying the column name representing the lag
(i.e., the difference between issue date and reference date).}

\item{min_refd}{the earliest reference date considered in the data}

\item{max_refd}{the latest reference date considered in the data}

\item{ref_lag}{An integer to indicate the reference lag}
}
\value{
df_new Data Frame with filled rows for missing lags
}
\description{
Re-index, fill na, make sure all reference date have enough rows for updates
}
