% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fips_to_geovalue}
\alias{fips_to_geovalue}
\title{Make sure data contains a \code{geo_value} field}
\usage{
fips_to_geovalue(df)
}
\arguments{
\item{df}{data.frame}
}
\description{
Make sure data contains a \code{geo_value} field
}
