% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_quantile_lasso}
\alias{fit_quantile_lasso}
\title{Retrieve or Train a Model}
\usage{
fit_quantile_lasso(
  x,
  y,
  tau,
  lambda,
  standardize = TRUE,
  intercept = TRUE,
  weights = NULL
)
}
\arguments{
\item{tau}{Quantile to be predicted (between 0 and 1).}

\item{model_path}{Path to the cached model file.}

\item{train_data}{Data frame containing the training data.}

\item{response}{Name of the response variable.}

\item{sqrt_max_raw}{Maximum raw value at square root level.}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{covariates-template}{\link{DESCRIPTION GOES HERE}}

\item{lp_solver-template}{\link{DESCRIPTION GOES HERE}}

\item{train_models-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
The trained or loaded model object.
}
\description{
This function retrieves a cached model if available or trains a new quantile regression model if necessary.
The trained model is saved with additional metadata attributes.
}
