% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_files_list}
\alias{get_files_list}
\title{List valid input files.}
\usage{
get_files_list(indicator, signal, params, sub_dir)
}
\arguments{
\item{sub_dir}{string specifying the indicator-specific directory within
the general input directory \code{params$input_dir}}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{params-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
List valid input files.
}
