% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_populous_counties}
\alias{get_populous_counties}
\title{Subset list of counties to those included in the 200 most populous in the US}
\usage{
get_populous_counties()
}
\description{
Requires the covidcast package, which is not installed by default.
}
