% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_report_date_range}
\alias{get_report_date_range}
\title{Find the earliest and latest issue dates needed for either training or testing.}
\usage{
get_report_date_range(params)
}
\arguments{
\item{params-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
With current logic, we always need to include data for model training (in
case cached models are not available for a "make_predictions"-only run and
we need to train new models).
}
\details{
We generate test and train data by applying the following filters:
\itemize{
\item Test data is data where report_date is in params$test_dates
(as a continuous filter, min(params$test_dates) <= report_date <= max(params$test_dates) )
\item Train data is data where report_date < training_end_date; and
training_start_date < target_date <= training_end_date
}

Train data doesn't have an explicit lower bound on report_date, but we can
derive one.

Since target_date = reference_date + params$ref_lag and report_date >=
reference_date, the requirement that training_start_date < target_date
also implies that issue date must be > training_start_date - params$ref_lag
}
