% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_training_date_range}
\alias{get_training_date_range}
\title{Get date range of data to use for training models}
\usage{
get_training_date_range(params)
}
\arguments{
\item{params-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Calculate training start and end dates based on user settings.
\code{training_start_date} is the minimum allowed target date when selecting
training data to use. \code{training_end_date} is the maximum allowed target
date and maximum allowed issue date.
}
\details{
Cases:
\enumerate{
\item We are training new models.
\item We are not training new models and cached models exist.
\item We are not training new models and cached models don't exist.
}

Sometimes we want to allow the user to specify an end date in
params that overrides the automatically-generated end date. This is
only relevant when the user requests to train new models.
}
