% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_weekofmonth}
\alias{get_weekofmonth}
\title{Get week of a month info according to a date}
\usage{
get_weekofmonth(date)
}
\arguments{
\item{date}{A Date object.}
}
\value{
An integer indicating the week number within the month.
}
\description{
Consider Epiweeks. Notice that
If there are 4, 5 or 6 weeks in total, the ith weeks is labeled as i
and the dates in the last week this month are actually in the same
week with the dates in the 1st week next month and those dates are
sparse. Thus, we assign the dates in the last week to the 1st week.
If there are 6 weeks in total, the 1st, 2nd, 3rd, 4th, 5th, 6th weeks
are labeled as c(1, 2, 3, 4, 5, 1) which means we will merge the first,
second and the last weeks together.
}
