% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_hyperparam}
\alias{handle_hyperparam}
\title{Handle Hyperparameter Selection Based on Test Lag Group}
\usage{
handle_hyperparam(param, test_lag_group)
}
\arguments{
\item{param}{A numeric value or a named list containing hyperparameters for different \code{test_lag_group}s.}

\item{test_lag_group}{A character or numeric value representing the test lag group.}
}
\value{
The selected hyperparameter value as a numeric.
}
\description{
This function selects the appropriate hyperparameter value based on whether
\code{param} is a named list or a single numeric value. If \code{param} is a list and
\code{test_lag_group} exists as a key, it returns the corresponding value. If not,
it falls back to \code{param$others}. If \code{param} has no names, it assumes a single
numeric value and returns it.
}
\examples{
# Case where param is a single numeric value
handle_hyperparam(0.1, 1) # Returns 0.1

# Case where param is a named list
param_list <- list("1" = 0.1, "2" = 0.2, "others" = 0.3)
handle_hyperparam(param_list, 1) # Returns 0.1
handle_hyperparam(param_list, 3) # Returns 0.3 (default "others")

}
