% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_dph.R
\docType{data}
\name{ma_dph}
\alias{ma_dph}
\title{Example Dataset: mydata}
\format{
A CSV file with 257381 rows and 5 variables:
}
\source{
COIVD-19 confirmed cases from ma-dph
}
\usage{
data(ma_dph)
}
\description{
This dataset contains sample data in CSV format.
}
\keyword{datasets}
