% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_key}
\alias{make_key}
\title{Generate key for identifying a value_type-signal combo}
\usage{
make_key(value_type, signal_suffix)
}
\arguments{
\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
If \code{signal_suffix} is not an empty string, concatenate the two arguments.
Otherwise, return only \code{value_type}.
}
