% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{params_element_exists_and_valid}
\alias{params_element_exists_and_valid}
\title{Check if an element in params exists and is not missing}
\usage{
params_element_exists_and_valid(params, key)
}
\arguments{
\item{key}{string indicating name of element within \code{params} to check}

\item{params-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Check if an element in params exists and is not missing
}
