% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_data}
\alias{read_data}
\title{Read a parquet file into a dataframe}
\usage{
read_data(input_file)
}
\arguments{
\item{input_file}{A string for the path to the input file}
}
\description{
Read a parquet file into a dataframe
}
