% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validity_checks}
\alias{validity_checks}
\title{Check input data for validity}
\usage{
validity_checks(
  df,
  value_type,
  num_col,
  denom_col,
  signal_suffixes,
  refd_col = "reference_date",
  lag_col = "lag",
  issued_col = "report_date"
)
}
\arguments{
\item{df}{A data.frame}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{num_col-template}{\link{DESCRIPTION GOES HERE}}

\item{denom_col-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffixes-template}{\link{DESCRIPTION GOES HERE}}

\item{refd_col-template}{\link{DESCRIPTION GOES HERE}}

\item{lag_col-template}{\link{DESCRIPTION GOES HERE}}

\item{issued_col-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
list of input dataframe augmented with lag column, if it
didn't already exist, and character vector of one or two value
column names, depending on requested \code{value_type}
}
\description{
Check input data for validity
}
