% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TAUS}
\alias{TAUS}
\alias{REF_LAG}
\alias{LAG_PAD}
\alias{TRAINING_DAYS}
\alias{LAMBDA}
\alias{GAMMA}
\alias{LP_SOLVER}
\alias{YITL}
\alias{SLOPE}
\alias{Y7DAV}
\alias{RESPONSE}
\alias{LOG_LAG}
\alias{WEEKDAYS_ABBR}
\alias{WEEK_ISSUES}
\alias{SQRT_SCALES}
\alias{TODAY}
\alias{TEST_LAG_GROUPS}
\alias{INDICATORS_AND_SIGNALS}
\title{Constants}
\format{
An object of class \code{numeric} of length 9.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 7.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{Date} of length 1.

An object of class \code{character} of length 19.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
TAUS

REF_LAG

LAG_PAD

TRAINING_DAYS

LAMBDA

GAMMA

LP_SOLVER

YITL

SLOPE

Y7DAV

RESPONSE

LOG_LAG

WEEKDAYS_ABBR

WEEK_ISSUES

SQRT_SCALES

TODAY

TEST_LAG_GROUPS

INDICATORS_AND_SIGNALS
}
\description{
A set of constants used in the model.
}
\details{
Constants

These constants are used for forecast revision.
}
\keyword{datasets}
