% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{cv_revision_forecast}
\alias{cv_revision_forecast}
\title{Cross-Validation for Forecast Revision}
\usage{
cv_revision_forecast(
  df,
  test_lag,
  taus = TAUS,
  smoothed_target = TRUE,
  lagged_term_list = NULL,
  params_list = NULL,
  temporal_resol = "daily",
  lambda_candidates = c(0.01, 0.1, 1),
  gamma_candidates = c(0.1, 1, 10),
  lag_pad_candidates = c(0, 1, 2, 3),
  lp_solver = LP_SOLVER,
  backend = SOLVER_BACKEND,
  geo = "ma",
  value_type = "count",
  model_save_dir = "./receiving",
  indicator = "testdata",
  signal = "",
  geo_level = "state",
  signal_suffix = "",
  training_end_date = "",
  training_days = 365,
  n_folds = 5
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{test_lag}{Numeric vector specifying the test lag(s).}

\item{taus}{Numeric vector of quantiles to estimate.}

\item{smoothed_target}{A logical value indicating whether the target variable should be smoothed.}

\item{lagged_term_list}{Optional list of lagged terms to be included in the model.}

\item{params_list}{Optional list of additional parameters for the model.}

\item{temporal_resol}{Character; either "daily" or "weekly" resolution.}

\item{lambda_candidates}{Numeric vector of lambda values to try.}

\item{gamma_candidates}{Numeric vector of gamma values to try.}

\item{lag_pad_candidates}{Numeric vector of lag padding values to try.}

\item{n_folds}{Integer, number of cross-validation folds.}

\item{train_models-template}{\link{DESCRIPTION GOES HERE}}

\item{make_predictions-template}{\link{DESCRIPTION GOES HERE}}

\item{model_save_dir-template}{\link{DESCRIPTION GOES HERE}}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{lp_solver-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{geo_level-template}{\link{DESCRIPTION GOES HERE}}

\item{geo-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{training_end_date-template}{\link{DESCRIPTION GOES HERE}}

\item{training_days-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
This function performs cross-validation to optimize WIS for forecast revision.
It trains forecast models for a given geographical location and a lag set.
It evaluates different combinations of lambda, gamma, and lag padding values
across multiple folds.
}
