% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Data Preprocessing Function}
\usage{
data_preprocessing(
  df,
  value_col,
  refd_col,
  lag_col,
  ref_lag,
  suffixes = c(""),
  lagged_term_list = NULL,
  value_type = "count",
  temporal_resol = "daily",
  smoothed = FALSE
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{value_col}{Character vector specifying the column(s) containing values.}

\item{refd_col}{Character specifying the reference column for merging operations.}

\item{lag_col}{Character specifying the column representing lag information.}

\item{ref_lag}{Integer specifying the reference lag value.}

\item{suffixes}{Character vector containing suffixes for value columns.}

\item{lagged_term_list}{Numeric vector specifying the list of lags to be considered.}

\item{value_type}{Character indicating the type of values ('count' or 'fraction').}

\item{temporal_resol}{Character specifying temporal resolution ('daily' or 'weekly').}

\item{smoothed}{Logical indicating whether smoothing should be applied.}
}
\description{
This function processes input data by handling missing values, computing lagged terms,
generating smoothed values, and transforming data for further analysis.
It supports both count and fraction data types and ensures consistency in column naming.
If two value column names are provided and the data type is "fraction", then the first
value column is treated as the column for numerators while the second column is treated as
the column for denominators.
}
