% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{DelphiRF}
\alias{DelphiRF}
\title{Revision Forecast for a Specific Location}
\usage{
DelphiRF(
  df,
  testing_start_date,
  taus = TAUS,
  test_lag_groups = NULL,
  smoothed_target = TRUE,
  lagged_term_list = NULL,
  params_list = NULL,
  extra_params = NULL,
  lambda = LAMBDA,
  gamma = GAMMA,
  lag_pad = LAG_PAD,
  temporal_resol = "daily",
  lp_solver = LP_SOLVER,
  backend = SOLVER_BACKEND,
  geo = "ma",
  value_type = "count",
  model_save_dir = "./receiving",
  indicator = "testdata",
  signal = "",
  geo_level = "state",
  signal_suffix = "",
  training_end_date = "",
  training_days = 365,
  train_models = TRUE,
  make_predictions = TRUE,
  onehot_weekdays = list(Mon = c("Mon"), Weekends = c("Sat", "Sun")),
  time_limit = NULL
)
}
\arguments{
\item{df}{A data frame containing historical data with \code{target_date} and \code{report_date} columns.}

\item{testing_start_date}{A character or Date object specifying the start date for testing.}

\item{taus}{A numeric vector of quantiles for probabilistic forecasting.}

\item{test_lag_groups}{A vector of test lag groups to process (default: \code{TEST_LAG_GROUPS}).}

\item{smoothed_target}{A logical value indicating whether the target variable should be smoothed.}

\item{lagged_term_list}{A list of lagged terms to be used as predictors.}

\item{params_list}{A list of model parameters for forecasting.}

\item{lag_pad}{A numeric or named list of lag padding values (default: \code{LAG_PAD}).}

\item{temporal_resol}{Character; either "daily" or "weekly" resolution.}

\item{train_models-template}{\link{DESCRIPTION GOES HERE}}

\item{make_predictions-template}{\link{DESCRIPTION GOES HERE}}

\item{model_save_dir-template}{\link{DESCRIPTION GOES HERE}}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{lp_solver-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{geo_level-template}{\link{DESCRIPTION GOES HERE}}

\item{geo-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{training_end_date-template}{\link{DESCRIPTION GOES HERE}}

\item{training_days-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
This function performs a forecasting operation using historical data, applying
hyper-parameters and lag-based adjustments for different test lag groups.
}
