% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{exponentiate_preds}
\alias{exponentiate_preds}
\title{Un-log predicted values}
\usage{
exponentiate_preds(test_data, taus)
}
\arguments{
\item{test_data}{dataframe with a column containing the prediction results of
each requested quantile. Each row represents an update with certain
(reference_date, report_date, location) combination.}

\item{taus-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Un-log predicted values
}
