% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_filename}
\alias{generate_filename}
\title{Construct filename for model with given parameters}
\usage{
generate_filename(
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  gamma,
  training_end_date,
  training_days = 365,
  geo = "",
  value_type = "",
  test_lag_group = "",
  tau = "",
  model_mode = TRUE,
  model_save_dir = file.path(tempdir(), "DelphiRF", "models")
)
}
\arguments{
\item{training_end_date}{End date assigned to the training data. High-level
forecasting functions use the latest input report date when this is empty.}

\item{training_days}{Number of days assigned to the training target-date
window. Functions receiving an unfiltered dataset apply this window;
\code{\link[=revision_forecast]{revision_forecast()}} expects pre-filtered input and records the value in
model cache filenames.}

\item{tau}{Decimal quantile to be predicted. Values must be between 0 and 1.}

\item{model_mode}{Boolean, indicates whether the file name is for a model.}

\item{model_save_dir}{Directory for model cache files. The default is a
session temporary directory. Supply an explicit path to keep models after
the R session ends.}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{geo-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{test_lag_group-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
Path to file containing model object.
}
\description{
Construct filename for model with given parameters
}
