% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_filename}
\alias{generate_filename}
\title{Construct filename for model with given parameters}
\usage{
generate_filename(
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  gamma,
  training_end_date,
  training_days = 365,
  geo = "",
  value_type = "",
  test_lag_group = "",
  tau = "",
  model_mode = TRUE
)
}
\arguments{
\item{tau}{Decimal quantile to be predicted. Values must be between 0 and 1.}

\item{model_mode}{Boolean, indicates whether the file name is for a model.}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{geo-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{test_lag_group-template}{\link{DESCRIPTION GOES HERE}}

\item{training_end_date-template}{\link{DESCRIPTION GOES HERE}}

\item{training_start_date-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
Path to file containing model object.
}
\description{
Construct filename for model with given parameters
}
