% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_prediction}
\alias{get_prediction}
\title{Fetch model and use to generate predictions/perform corrections}
\usage{
get_prediction(
  test_data,
  taus,
  covariates,
  response,
  obj,
  make_evaluation = TRUE
)
}
\arguments{
\item{taus}{vector of numeric quantiles}

\item{covariates}{vector of strings for the name of covariates}

\item{obj}{pre-trained model}

\item{make_evaluation}{Logical}

\item{test_data-template}{\link{DESCRIPTION GOES HERE}}

\item{response-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Fetch model and use to generate predictions/perform corrections
}
