% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{revision_forecast}
\alias{revision_forecast}
\title{Revision Forecast Based on Training and Testing Data}
\usage{
revision_forecast(
  train_data,
  test_data,
  taus,
  smoothed_target = TRUE,
  lagged_term_list = NULL,
  params_list = NULL,
  extra_params = NULL,
  temporal_resol = "daily",
  lambda = 0.1,
  gamma = 0.1,
  lp_solver = LP_SOLVER,
  test_lag_group = "",
  geo = "ma",
  value_type = "count",
  model_save_dir = "./receiving",
  indicator = "testdata",
  signal = "",
  geo_level = "state",
  signal_suffix = "",
  training_end_date = "",
  training_days = 365,
  train_models = TRUE,
  make_predictions = TRUE,
  onehot_weekdays = list(Mon = c("Mon"), Weekends = c("Sat", "Sun")),
  time_limit = NULL,
  backend = SOLVER_BACKEND
)
}
\arguments{
\item{train_data}{A data frame containing the training dataset.}

\item{test_data}{A data frame containing the testing dataset.}

\item{taus}{A numeric vector specifying the quantiles for prediction intervals.}

\item{smoothed_target}{A logical value indicating whether the target variable should be smoothed.}

\item{lagged_term_list}{A numeric vector specifying the lag values for which reported
values on the reference date are considered in model training.}

\item{params_list}{A list of model parameters used for training and prediction.}

\item{train_models-template}{\link{DESCRIPTION GOES HERE}}

\item{make_predictions-template}{\link{DESCRIPTION GOES HERE}}

\item{model_save_dir-template}{\link{DESCRIPTION GOES HERE}}

\item{indicator-template}{\link{DESCRIPTION GOES HERE}}

\item{signal-template}{\link{DESCRIPTION GOES HERE}}

\item{lp_solver-template}{\link{DESCRIPTION GOES HERE}}

\item{geo_level-template}{\link{DESCRIPTION GOES HERE}}

\item{geo-template}{\link{DESCRIPTION GOES HERE}}

\item{signal_suffix-template}{\link{DESCRIPTION GOES HERE}}

\item{lambda-template}{\link{DESCRIPTION GOES HERE}}

\item{gamma-template}{\link{DESCRIPTION GOES HERE}}

\item{value_type-template}{\link{DESCRIPTION GOES HERE}}

\item{test_lag_group-template}{\link{DESCRIPTION GOES HERE}}

\item{training_end_date-template}{\link{DESCRIPTION GOES HERE}}

\item{training_days-template}{\link{DESCRIPTION GOES HERE}}
}
\value{
A list containing:
\item{test_data}{The test dataset with added forecast evaluations.}
\item{coefs}{Model coefficients from the trained model.}
}
\description{
This function trains a forecast model for a given geographical location
and lag set. It assumes that the input training and testing data are
already well-filtered. It also evaluates forecast accuracy using the
weighted interval score (WIS).
}
