% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{training_days_check}
\alias{training_days_check}
\title{Check available training days}
\usage{
training_days_check(report_date, training_days)
}
\arguments{
\item{report_date}{contents of input data's \code{report_date} column}

\item{training_days-template}{\link{DESCRIPTION GOES HERE}}
}
\description{
Check available training days
}
